+++ src/Completion/Tool/Tool.php
@@ -7,7 +7,7 @@ namespace ByCerfrance\LlmApiLib\Completion\Tool;
 use Closure;
 use Override;
 
-readonly class Tool implements ToolInterface
+readonly class Tool extends AbstractTool
 {
     private Closure $callback;
 
@@ -18,48 +18,18 @@ readonly class Tool implements ToolInterface
      * @param callable $callback Callback to execute when tool is called
      */
     public function __construct(
-        private string $name,
-        private string $description,
-        private array $parameters,
+        string $name,
+        string $description,
+        array $parameters,
         callable $callback,
     ) {
+        parent::__construct($name, $description, $parameters);
         $this->callback = $callback(...);
-    }
-
-    #[Override]
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    #[Override]
-    public function getDescription(): string
-    {
-        return $this->description;
-    }
-
-    #[Override]
-    public function getParameters(): array
-    {
-        return $this->parameters;
     }
 
     #[Override]
     public function execute(array $arguments): mixed
     {
         return ($this->callback)($arguments);
-    }
-
-    #[Override]
-    public function jsonSerialize(): array
-    {
-        return [
-            'type' => 'function',
-            'function' => [
-                'name' => $this->name,
-                'description' => $this->description,
-                'parameters' => (object)$this->parameters,
-            ],
-        ];
     }
 }